+++ app/src/main/java/org/simple/clinic/editpatient/EditPatientEvent.kt
@@ -3,6 +3,7 @@ package org.simple.clinic.editpatient
 import org.simple.clinic.newentry.country.InputFields
 import org.simple.clinic.patient.Gender
 import org.simple.clinic.patient.businessid.BusinessId
+import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.widgets.UiEvent
 
 sealed class EditPatientEvent : UiEvent
@@ -75,4 +76,12 @@ data class ColonyOrVillagesFetched(val colonyOrVillages: List<String>) : EditPat
 
 object AddNHIDButtonClicked : EditPatientEvent() {
   override val analyticsName = "Edit Patient Entry:Add NHID Button Clicked"
+}
+
+object AddBpPassportButtonClicked : EditPatientEvent() {
+  override val analyticsName = "Edit Patient Entry:Add Bp Passport Clicked"
+}
+
+data class BpPassportAdded(val identifier: List<Identifier>) : EditPatientEvent() {
+  override val analyticsName: String = "Edit Patient Entry:Bp Passport Added"
 }