+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Bump facebook.soloader to 0.10.3
 - Bump Mobius to v1.5.7
 - [In Progress: 11 Jan 2022] Implement adding NHID in `PatientEditScreen`
+- [In Progress: 17 Jan 2022] Implement adding scan support for BpPassport in `EditPatientScreen`
 
 ## On Demo